fix: permit removal of manually removed services
A community user manually removed one of his service file definitions and he then discovered that
the node manager would not allow the service to be removed from its registry.

These changes permit the manual removal of the service definition and also the data and log
directories. If they have been removed, we just proceed to remove the service from the node
manager's registry. This solution has only been tested for Systemd-based services. It would need to
be expanded and tested for other platforms and service managers.

It would have been possible to submit a PR for the service manager crate to behave in an idempotent
way and not error if the service file had been removed, but it seems correct for the service manager
to error in that case. We can handle the error in our application.

+++ sn_node_manager/src/lib.rs
@@ -39,6 +39,7 @@ use colored::Colorize;
 use semver::Version;
 use sn_service_management::{
     control::ServiceControl,
+    error::Error as ServiceError,
     rpc::{RpcActions, RpcClient},
     NodeRegistry, NodeServiceData, ServiceStateActions, ServiceStatus, UpgradeOptions,
     UpgradeResult,
@@ -180,11 +181,29 @@ impl<T: ServiceStateActions + Send> ServiceManager<T> {
             );
         }
 
-        self.service_control.uninstall(&self.service.name())?;
+        match self.service_control.uninstall(&self.service.name()) {
+            Ok(()) => {}
+            Err(e) => match e {
+                ServiceError::ServiceRemovedManually(name) => {
+                    // The user has deleted the service definition file, which the service manager
+                    // crate treats as an error. We then return our own error type, which allows us
+                    // to handle it here and just proceed with removing the service from the
+                    // registry.
+                    println!("The user appears to have removed the {name} service manually");
+                }
+                _ => return Err(e.into()),
+            },
+        }
 
         if !keep_directories {
-            std::fs::remove_dir_all(self.service.data_dir_path())?;
-            std::fs::remove_dir_all(self.service.log_dir_path())?;
+            // It's possible the user deleted either of these directories manually.
+            // We can just proceed with removing the service from the registry.
+            if self.service.data_dir_path().exists() {
+                std::fs::remove_dir_all(self.service.data_dir_path())?;
+            }
+            if self.service.log_dir_path().exists() {
+                std::fs::remove_dir_all(self.service.log_dir_path())?;
+            }
         }
 
         self.service.on_remove();
